+++ Funcky.Test/Monads/EitherTest.cs
@@ -93,7 +93,7 @@ namespace Funcky.Test.Monads
             var hasLeft = false;
             var hasRight = false;
 
-            value.Match(
+            value.Switch(
                 left: _ => Execute(() => hasLeft = true),
                 right: _ => Execute(() => hasRight = true));
 
@@ -109,7 +109,7 @@ namespace Funcky.Test.Monads
             var hasLeft = false;
             var hasRight = false;
 
-            value.Match(
+            value.Switch(
                 left: _ => Execute(() => hasLeft = true),
                 right: _ => Execute(() => hasRight = true));